Move plugin package up
