+++ pipelex/cogt/config_cogt.py
@@ -6,15 +6,15 @@ from pipelex.cogt.imgg.imgg_handle import ImggHandle
 from pipelex.cogt.imgg.imgg_job_components import ImggJobConfig, ImggJobParams, ImggJobParamsDefaults
 from pipelex.cogt.llm.llm_job_components import LLMJobConfig
 from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
-from pipelex.cogt.plugin.anthropic.anthropic_config import AnthropicConfig
-from pipelex.cogt.plugin.bedrock.bedrock_config import BedrockConfig
-from pipelex.cogt.plugin.mistral.mistral_config import MistralConfig
-from pipelex.cogt.plugin.openai.azure_openai_config import AzureOpenAIConfig
-from pipelex.cogt.plugin.openai.custom_endpoint_config import CustomEndpointConfig
-from pipelex.cogt.plugin.openai.openai_config import OpenAIOpenAIConfig
-from pipelex.cogt.plugin.openai.perplexity_config import PerplexityConfig
-from pipelex.cogt.plugin.openai.vertexai_config import VertexAIConfig
-from pipelex.cogt.plugin.openai.xai_config import XaiConfig
+from pipelex.plugin.anthropic.anthropic_config import AnthropicConfig
+from pipelex.plugin.bedrock.bedrock_config import BedrockConfig
+from pipelex.plugin.mistral.mistral_config import MistralConfig
+from pipelex.plugin.openai.azure_openai_config import AzureOpenAIConfig
+from pipelex.plugin.openai.custom_endpoint_config import CustomEndpointConfig
+from pipelex.plugin.openai.openai_config import OpenAIOpenAIConfig
+from pipelex.plugin.openai.perplexity_config import PerplexityConfig
+from pipelex.plugin.openai.vertexai_config import VertexAIConfig
+from pipelex.plugin.openai.xai_config import XaiConfig
 from pipelex.tools.config.models import ConfigModel
