+++ pipelex/plugin/bedrock/__init__.py
