+++ pipelex/plugin/bedrock/bedrock_factory.py
@@ -2,11 +2,11 @@ from pipelex import log
 from pipelex.cogt.exceptions import LLMCapabilityError, PromptImageFormatError
 from pipelex.cogt.image.prompt_image import PromptImageBytes
 from pipelex.cogt.llm.llm_job import LLMJob
-from pipelex.cogt.plugin.bedrock.bedrock_client_protocol import BedrockClientProtocol
-from pipelex.cogt.plugin.bedrock.bedrock_config import BedrockClientMethod
-from pipelex.cogt.plugin.bedrock.bedrock_message import BedrockContentItem, BedrockImage, BedrockMessage, BedrockSource, ImageFormat
 from pipelex.config import get_config
 from pipelex.hub import get_secrets_provider
+from pipelex.plugin.bedrock.bedrock_client_protocol import BedrockClientProtocol
+from pipelex.plugin.bedrock.bedrock_config import BedrockClientMethod
+from pipelex.plugin.bedrock.bedrock_message import BedrockContentItem, BedrockImage, BedrockMessage, BedrockSource, ImageFormat
 
 
 class BedrockFactory:
@@ -23,11 +23,11 @@ class BedrockFactory:
         log.verbose(f"Using '{client_method}' for BedrockClient")
         match client_method:
             case BedrockClientMethod.AIBOTO3:
-                from pipelex.cogt.plugin.bedrock.bedrock_client_aioboto3 import BedrockClientAioboto3
+                from pipelex.plugin.bedrock.bedrock_client_aioboto3 import BedrockClientAioboto3
 
                 bedrock_async_client = BedrockClientAioboto3(aws_region=aws_region)
             case BedrockClientMethod.BOTO3:
-                from pipelex.cogt.plugin.bedrock.bedrock_client_boto3 import BedrockClientBoto3
+                from pipelex.plugin.bedrock.bedrock_client_boto3 import BedrockClientBoto3
 
                 bedrock_async_client = BedrockClientBoto3(aws_region=aws_region)
