+++ pipelex/plugin/openai/openai_imgg_factory.py
