+++ pipelex/plugin/openai/openai_llm_worker.py
@@ -15,7 +15,7 @@ from pipelex.cogt.llm.llm_models.llm_engine import LLMEngine
 from pipelex.cogt.llm.llm_models.llm_family import LLMFamily
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
-from pipelex.cogt.plugin.openai.openai_factory import OpenAIFactory
+from pipelex.plugin.openai.openai_factory import OpenAIFactory
 from pipelex.tools.typing.pydantic_utils import BaseModelTypeVar
