+++ tests/pipelex/cogt/synchro/llm_providers/test_mistral.py
@@ -4,7 +4,7 @@ from rich.console import Console
 from rich.table import Table
 
 from pipelex import pretty_print
-from pipelex.cogt.plugin.mistral.mistral_llms import list_mistral_models
+from pipelex.plugin.mistral.mistral_llms import list_mistral_models
 
 
 # make t VERBOSE=2 TEST=TestMistral
